Use uint64 type for entity IDs in the models
It's very possible for this ID to be much larger than an int32,
so we need to switch the type of this ID to be safe.

+++ backend/models/example/resolver.go
@@ -23,7 +23,7 @@ func (r *queryResolver) Streams(ctx context.Context) ([]models.Stream, error) {
 func (r *queryResolver) Stream(ctx context.Context, streamID int) (models.Stream, error) {
 	panic("not implemented")
 }
-func (r *queryResolver) Entity(ctx context.Context, streamID int, entityID int) (models.Entity, error) {
+func (r *queryResolver) Entity(ctx context.Context, streamID int, entityID uint64) (models.Entity, error) {
 	panic("not implemented")
 }
